アニメーション導入途中

+++ Senzan/SourceCode/Game/01_GameObject/00_MeshObject/00_Character/02_Boss/BossMoveState/BossMoveState.h
@@ -61,7 +61,7 @@ public:
 private:
 	//現在のボスの回転度を確認する.
 	float m_RotationAngle;
-	float m_RotationSpeed;
+	double m_RotationSpeed;
 	float m_rotationDirection;	//1.0f:順方向, -1.0:逆方向.
 
 	int							m_AnimNo;	//アニメーション番号.
